refactor(components): split group abilities
